renaming the mailer util
