+++ src/auth-service/config/mailer.config.js
